Delay inserting the template into the document until the first digest cycle.
Avoids issues with Angular compiling the element a second time when the document is parsed. When double-compiled, the ng-ifs do not function properly and the notify bar is displayed with no message. This issue is most likely encountered if the service is configured in a run block when the app is loading.